Changed file structure
