+++ variables.js
@@ -1,0 +1,15 @@
+export const variables =
+    [
+        {
+            name: 'The last received chat message',
+            variableId: 'lastMessage'
+        },
+        {
+            name: 'The last received chat id',
+            variableId: 'lastChatId'
+        },
+        {
+            name: 'The last received chat timestamp',
+            variableId: 'lastTimestamp'
+        },
+    ]
